Dashboard page refactor
