+++ multi-manager-dashboard/src/components/SnapshotCard/SnapshotsCardArea.tsx
