+++ multi-manager-dashboard/src/components/SnapshotCard/SnapshotsDeltas.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { Area, AreaChart, CartesianGrid, ResponsiveContainer, Tooltip, XAxis, YAxis } from 'recharts';
-import { ISnapshot_Delta } from '../redux/slices/vaultsSlice';
+import { ISnapshot_Delta } from '../../redux/slices/vaultsSlice';
 
 type Props = {
   deltas: ISnapshot_Delta;
