--- multi-manager-dashboard/src/pages/TransactionsPage.tsx
@@ -1,13 +1,0 @@
-import React from 'react'
-import { useParams } from 'react-router-dom';
-
-
-const TransactionsPage = () => {
-    return (
-        <div className='flex h-full justify-center'>
-            <div className='mt-16 text-4xl text-gray-400'>Coming Soon</div>
-        </div>
-    )
-}
-
-export default TransactionsPage
